refactored, added itemfilterfragment
