+++ app/src/main/java/com/example/cmput301project/Fragments/ItemFiltersFragment.java
@@ -1,4 +1,4 @@
-package com.example.cmput301project;
+package com.example.cmput301project.Fragments;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -9,15 +9,22 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 
+import com.example.cmput301project.ItemClasses.ItemFilter;
+import com.example.cmput301project.R;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class ItemFiltersFragment extends DialogFragment {
-    private TextView title;
-    private EditText date;
+    private EditText editFromDate;
+    private EditText editToDate;
+    private EditText editKeywords;
+    private EditText editMakes;
     private OnFragmentInteractionListener listener;
 
     public ItemFiltersFragment(){
@@ -36,13 +43,16 @@ public class ItemFiltersFragment extends DialogFragment {
     }
 
     public interface OnFragmentInteractionListener {
-        void onFiltersPressed();
+        void onFiltersSaved(ItemFilter itemFilter);
     }
 
     @NonNull
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
         View view = LayoutInflater.from(getActivity()).inflate(R.layout.item_filters, null);
+
+        editFromDate = view.findViewById(R.id.from_date_edit_text);
+        editToDate = view.findViewById(R.id.to_date_edit_text);
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
 
@@ -57,8 +67,24 @@ public class ItemFiltersFragment extends DialogFragment {
                 okButton.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View view) {
+                        ItemFilter itemFilter = new ItemFilter();
 
+                        String fromDateString = editFromDate.getText().toString().trim();
+                        String toDateString = editToDate.getText().toString().trim();
 
+                        if (!fromDateString.isEmpty() && !toDateString.isEmpty()) {
+                            SimpleDateFormat dateFormat = new SimpleDateFormat("MM/dd/yyyy");
+                            try {
+                                Date fromDate = dateFormat.parse(fromDateString);
+                                Date toDate = dateFormat.parse(toDateString);
+                                itemFilter.setFrom(fromDate);
+                                itemFilter.setTo(toDate);
+                                itemFilter.setFilterDate(true);
+                            } catch (Exception e) {
+                                e.printStackTrace();
+                            }
+                        }
+                        listener.onFiltersSaved(itemFilter);
                     }
                 });
             }
@@ -66,4 +92,5 @@ public class ItemFiltersFragment extends DialogFragment {
 
         return dialog;
     }
+
 }
