+++ app/src/main/java/com/example/cmput301project/ItemClasses/ItemList.java
@@ -1,0 +1,7 @@
+package com.example.cmput301project.ItemClasses;
+
+import java.util.ArrayList;
+
+public class ItemList {
+    private ArrayList<Item> items;
+}
