+++ app/src/main/java/com/example/cmput301project/ItemClasses/Photograph.java
@@ -1,0 +1,17 @@
+package com.example.cmput301project.ItemClasses;
+
+public class Photograph {
+    private String path;
+
+    public Photograph(String path) {
+        this.path = path;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+}
