+++ app/src/main/java/com/example/cmput301project/ItemClasses/Tag.java
@@ -1,0 +1,17 @@
+package com.example.cmput301project.ItemClasses;
+
+public class Tag {
+    private String name;
+
+    public Tag(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
