+++ app/src/main/java/com/example/cmput301project/MainActivity.java
@@ -9,10 +9,18 @@ import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import com.example.cmput301project.Fragments.AddItemFragment;
+import com.example.cmput301project.Fragments.ItemFiltersFragment;
+import com.example.cmput301project.Fragments.ViewItemFragment;
+import com.example.cmput301project.ItemClasses.Item;
+import com.example.cmput301project.ItemClasses.ItemAdapter;
+import com.example.cmput301project.ItemClasses.ItemFilter;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class MainActivity extends AppCompatActivity implements AddItemFragment.OnFragmentInteractionListener, ViewItemFragment.OnFragmentInteractionListener, ItemFiltersFragment.OnFragmentInteractionListener {
     private ArrayList<Item> items;
@@ -21,6 +29,7 @@ public class MainActivity extends AppCompatActivity implements AddItemFragment.O
     private Button filtersButton;
     private Double totalCost;
     private ArrayAdapter<Item> itemAdapter;
+    private ArrayAdapter<Item> filteredItemAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,7 +41,7 @@ public class MainActivity extends AppCompatActivity implements AddItemFragment.O
         filtersButton = findViewById(R.id.filter_items_button);
 
         totalCost = 0.00; //initialize costs to 0 on startup
-        itemAdapter = new ItemList(this, items);
+        itemAdapter = new ItemAdapter(this, items);
         itemsView.setAdapter(itemAdapter);
         final FloatingActionButton addButton = findViewById(R.id.add_item_button);
         addButton.setOnClickListener(v -> {
@@ -74,6 +83,23 @@ public class MainActivity extends AppCompatActivity implements AddItemFragment.O
         itemAdapter.notifyDataSetChanged();
     }
     @Override
-    public void onFiltersPressed() {}
+    public void onFiltersSaved(ItemFilter itemFilter) {
+        ArrayList<Item> filteredItems = new ArrayList<>();
+    }
+
+    public ArrayList<Item> filterByDate(Date from, Date to) {
+        List<Item> filteredItems = items.stream()
+                .filter(item -> item.getPurchaseDate().after(from) && item.getPurchaseDate().before(to)) // Replace this condition with your filtering criteria
+                .collect(Collectors.toList());
+        ArrayList<Item> filteredItemsArrayList = new ArrayList<>(filteredItems);
+        return filteredItemsArrayList;
+    }
+    public ArrayList<Item> filterByKeywords() {
+        return items;
+    }
+
+    public ArrayList<Item> filterByMake() {
+        return items;
+    }
 
 }
