--- app/src/main/java/com/example/cmput301project/Photograph.java
@@ -1,17 +1,0 @@
-package com.example.cmput301project;
-
-public class Photograph {
-    private String path;
-
-    public Photograph(String path) {
-        this.path = path;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-}
