--- app/src/main/java/com/example/cmput301project/Tag.java
@@ -1,17 +1,0 @@
-package com.example.cmput301project;
-
-public class Tag {
-    private String name;
-
-    public Tag(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
