+++ app/src/main/res/layout/activity_main.xml
@@ -14,6 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
+        android:gravity="center"
         tools:context=".MainActivity">
         <TextView
             android:layout_width="wrap_content"
