+++ app/src/main/res/layout/item_filters.xml
@@ -5,29 +5,37 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
     <TextView
-        android:id="@+id/item_filters_title"
         android:layout_width="match_parent"
         android:layout_height="75dp"
         android:text="Filters"
         android:gravity="center"
-        android:textSize="25dp"/>
+        android:textSize="25sp"/>
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:gravity="center"
         android:orientation="horizontal">
+        <EditText
+            android:id="@+id/from_date_edit_text"
+            android:inputType="date"
+            android:layout_width="100dp"
+            android:layout_height="wrap_content"
+            android:textSize="20sp"
+            android:hint="Start Date"
+            android:autofillHints="name">
+        </EditText>
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
-            android:text="Date Range"
-            android:textSize="25dp"
-            android:padding="5dp"
-            android:layout_marginRight="10dp"/>
+            android:text=" to "
+            android:textSize="25sp"/>
         <EditText
-            android:id="@+id/date_edit_text"
-            android:layout_width="match_parent"
+            android:inputType="date"
+            android:id="@+id/to_date_edit_text"
+            android:layout_width="100dp"
             android:layout_height="wrap_content"
             android:textSize="20sp"
-            android:hint="Date">
+            android:hint="End Date">
         </EditText>
     </LinearLayout>
 </LinearLayout>
